# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for what went wrong",
        "title": "What Went Wrong API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/excuse-templates": {
            "get": {
                "description": "Get excuse templates. Can filter by pack_id. Premium users can access all. Free users restricted from premium packs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuse-templates"
                ],
                "summary": "List excuse templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pack ID to filter",
                        "name": "pack_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetExcuseTemplatesResponse"
                        }
                    }
                }
            }
        },
        "/excuse-templates/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "excuse-templates"
                ],
                "summary": "Get template details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExcuseTemplateResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.TemplateNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.TemplateInternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/goals": {
            "get": {
                "description": "Get all goals for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "List goals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetGoalsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalFetchErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalFetchErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new goal. Checks for plan limits.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Create goal",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden if max goals reached",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalLimitReachedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalCreateErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/goals/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Delete goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalNotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalDeleteErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Update goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalNotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoalUpdateErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateGoalRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "notificationEnabled": {
                    "type": "boolean",
                    "example": true
                },
                "notificationTime": {
                    "type": "string",
                    "example": "20:00"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "本を10ページ読む"
                }
            }
        },
        "handlers.CreateGoalResponse": {
            "type": "object",
            "properties": {
                "goal": {
                    "$ref": "#/definitions/handlers.GoalResponse"
                }
            }
        },
...
        "handlers.ExcuseTemplateResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "excuseText": {
                    "type": "string",
                    "example": "宿題を犬に食べられました。"
                },
                "id": {
                    "type": "string",
                    "example": "template_123"
                },
                "isPremium": {
                    "type": "boolean",
                    "example": false
                },
                "packId": {
                    "type": "string",
                    "example": "pack_abc"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "面白い",
                        "定番"
                    ]
                }
            }
        },
...
        "handlers.GetExcuseTemplatesResponse": {
            "type": "object",
            "properties": {
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ExcuseTemplateResponse"
                    }
                }
            }
        },
...
        "handlers.GetGoalsResponse": {
            "type": "object",
            "properties": {
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.GoalResponse"
                    }
                }
            }
        },
...
        "handlers.GoalCreateErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "目標の作成に失敗しました"
                }
            }
        },
        "handlers.GoalDeleteErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "目標の削除に失敗しました"
                }
            }
        },
        "handlers.GoalFetchErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "目標の取得に失敗しました"
                }
            }
        },
        "handlers.GoalLimitReachedResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "プランの目標作成数上限に達しました"
                }
            }
        },
        "handlers.GoalNotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "目標が見つかりません"
                }
            }
        },
        "handlers.GoalResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "notificationEnabled": {
                    "type": "boolean",
                    "example": true
                },
                "notificationTime": {
                    "type": "string",
                    "example": "20:00"
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "本を10ページ読む"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "handlers.GoalUpdateErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "目標の更新に失敗しました"
                }
            }
        },
        "handlers.GoalValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "入力内容が正しくありません"
                }
            }
        },
...
        "handlers.TemplateInternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "テンプレートの取得に失敗しました"
                }
            }
        },
        "handlers.TemplateNotFoundResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "テンプレートが見つかりません"
                }
            }
        },
...
        "handlers.UpdateGoalRequest": {
            "type": "object",
            "properties": {
                "notificationEnabled": {
                    "type": "boolean",
                    "example": false
                },
                "notificationTime": {
                    "type": "string",
                    "example": "21:00"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "本を20ページ読む"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}